Refactor XCom page and XCom columns.
