+++ airflow/ui/src/components/ui/index.ts
@@ -35,3 +35,4 @@ export * from "./Status";
 export * from "./Button";
 export * from "./Toaster";
 export * from "./Breadcrumb";
+export * from "./Clipboard";
